test: 유효성 검사 테스트 분리
