+++ src/test/java/racingcar/ApplicationTest.java
@@ -41,51 +41,6 @@ class ApplicationTest extends NsTest {
         );
     }
 
-    @Test
-    @DisplayName("쉼표(,)로 구분했을 때 자동차 이름이 존재해야 한다")
-    void 자동차이름_존재하지_않을때() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi, ,", "1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Test
-    @DisplayName("자동차의 이름은 1자 이상 5자 이하여야 한다")
-    void 자동차이름길이() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobijjang,juni", "1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Test
-    @DisplayName("자동차의 이름은 같을 수 없다")
-    void 자동차이름동일여부() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,pobi", "1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Test
-    @DisplayName("시도 횟수는 숫자로 이루어져야 한다")
-    void 시도횟수_형식() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,juni", "aa"))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Test
-    @DisplayName("시도 횟수는 필수로 입력되어야 한다")
-    void 시도횟수_입력여부() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,juni", "\n"))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
     @Override
     public void runMain() {
         Application.main(new String[]{});
